+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v1.2.1 - [Unreleased]
  * Allow set custom type name and description for mutations `input` argument
  * Added `plural` in constraint violation error for pluralizing the violation message
  * Fixed check only forbidden type if the operation does not have any endpoint configured
+ * Added behat step to verify GraphQL error code
  * minors bugs fixed
   
 v1.2.0 - 2018-07-09